fix some bugs under 12/6-10

+++ 12/9.c
@@ -1,12 +1,13 @@
 #include <stdio.h>
 #define PI 3.14159
-int main() {
-	float r;
-    float c, s;
-	scanf("%g", &r);
+int main()
+{
+    double r;
+    double c, s;
+    scanf("%lf", &r);
     c = 2 * PI * r;
-    s = PI * r *r;
-    printf("%.3f\n",c);
-    printf("%.3f\n",s);
+    s = PI * r * r;
+    printf("%.3lf\n", c);
+    printf("%.3lf\n", s);
     return 0;
 }
